feat: разделяет PlaceOffer на компоненты
