+++ src/components/place-offer/offer-mark/offer-mark.tsx
@@ -1,0 +1,7 @@
+export default function OfferMark(): React.JSX.Element {
+  return(
+    <div className="offer__mark">
+      <span>Premium</span>
+    </div>
+  );
+}
