+++ src/components/reviews-section/reviews-section.tsx
@@ -1,7 +1,7 @@
 import { Reviews } from '../../types/reviews';
 
-import ReviewForm from '../reviews-form/reviews-form';
-import ReviewsList from '../reviews-list/reviews-list';
+import ReviewForm from './reviews-form/reviews-form';
+import ReviewsList from './reviews-list/reviews-list';
 
 type ReviewsSectionProps = {
   offerId: string;
